consolidated items, accounts and transactions into single domain model and moved all models into domainsmodels.py
