+++ purch/domains/finance/service.py
@@ -1,0 +1,5 @@
+from purch.domains.finance.repository import FinanceRepository
+
+
+class FinanceService:
+    finance_repo = FinanceRepository()
