+++ purch/domains/user/service.py
@@ -1,0 +1,5 @@
+from purch.domains.user.repository import UserRepository
+
+class UserService():
+    user_repo = UserRepository()
+
